Вернул один бин на место, переписал парсер.
